Add each paragraph as single component
