+++ src/paragraphs/BusinessYearParagraph.js
@@ -1,0 +1,14 @@
+import React from 'react'
+import Paragraph from '../Paragraph'
+
+const BusinessYearParagraph = () => {
+  return (
+    <Paragraph
+      id="businessYear"
+      headline="§ 2 Geschäftsjahr"
+      text="Geschäftsjahr ist das Kalenderjahr."
+    />
+  )
+}
+
+export default BusinessYearParagraph
