+++ src/paragraphs/MemberFeeParagraph.js
@@ -1,0 +1,18 @@
+import React from 'react'
+import Paragraph from '../Paragraph'
+
+const MemberFeeParagraph = ({ answers }) => {
+  return (
+    <Paragraph
+      id="memberFee"
+      headline="§ 8 Mitgliedsbeiträge"
+      text={
+        answers.memberFee === 'yesMemberFee'
+          ? 'Von den Mitgliedern werden Beiträge erhoben. Die Höhe der Beiträge und deren Fälligkeit bestimmt die Mitgliederversammlung.'
+          : 'Mitgliedsbeiträge werden nicht erhoben.'
+      }
+    />
+  )
+}
+
+export default MemberFeeParagraph
