+++ src/paragraphs/NonProfitParagraph.js
@@ -1,0 +1,15 @@
+import React from 'react'
+import Paragraph from '../Paragraph'
+
+const NonProfitParagraph = () => {
+  return (
+    <Paragraph
+      id="nonProfit"
+      headline="§ 4 Gemeinnützigkeit"
+      text="Der Verein ist selbstlos tätig. Er verfolgt nicht in erster Linie
+      eigenwirtschaftliche Zwecke."
+    />
+  )
+}
+
+export default NonProfitParagraph
