+++ src/paragraphs/UseOfFundsParagraph.js
@@ -1,0 +1,16 @@
+import React from 'react'
+import Paragraph from '../Paragraph'
+
+const UseOfFundsParagraph = () => {
+  return (
+    <Paragraph
+      id="useOfFunds"
+      headline="§ 5 Mittelverwendung"
+      text="Mittel des Vereins dürfen nur für die satzungsmäßigen Zwecke verwendet
+    werden. Die Mitglieder erhalten keine Zuwendungen aus Mitteln des
+    Vereins."
+    />
+  )
+}
+
+export default UseOfFundsParagraph
